backend/requirements.txt
# Data handling and analysis
pandas==2.2.1
numpy==1.26.4
yfinance==0.2.36

# Data visualization
matplotlib==3.8.3
seaborn==0.13.2

# API and data fetching
alpha_vantage==2.3.1
requests==2.31.0

# Date handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.1